dashboard/requirements.txt
streamlit
Pillow
numpy
pandas
plotly
openpyxl